feat: Application 로직을 JPQL 로 변경하기

+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -160,7 +160,7 @@ class BookServiceTest @Autowired constructor(
     
   }
 
-  private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+  private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
     assertThat(results.first {result -> result.type == type}.count).isEqualTo(count)
   }
